(#10) PlayerCollision: split into client and server
